+++ apps/dsp-app/src/app/project/collaboration/select-group/select-group.component.ts
@@ -16,7 +16,7 @@ import {
 import { ApplicationStateService } from '@dasch-swiss/vre/shared/app-state-service';
 import { DspApiConnectionToken } from '@dasch-swiss/vre/shared/app-config';
 import { AppErrorHandler } from '@dasch-swiss/vre/shared/app-error-handler';
-import { AutocompleteItem } from '@dsp-app/src/app/workspace/search/advanced-search/resource-and-property-selection/search-select-property/specify-property-value/operator';
+import { AutocompleteItem } from '@dsp-app/src/app/workspace/search/operator';
 
 @Component({
     selector: 'app-select-group',